+++ spec/rubocop/cop/lint/useless_assignment_spec.rb
@@ -4,293 +4,193 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
   subject(:cop) { described_class.new }
 
   context 'when a variable is assigned and unreferenced in a method' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         class SomeClass
           foo = 1
           puts foo
           def some_method
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(5)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned and unreferenced ' \
           'in a singleton method defined with self keyword' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         class SomeClass
           foo = 1
           puts foo
           def self.some_method
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(5)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced ' \
           'in a singleton method defined with variable name' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do
           foo = 1
           puts foo
           instance = Object.new
           def instance.some_method
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced in a class' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do
           foo = 1
           puts foo
           class SomeClass
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(5)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced in a class ' \
           'subclassing another class stored in local variable' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do
           foo = 1
           puts foo
           array_class = Array
           class SomeClass < array_class
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced ' \
           'in a singleton class' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do
           foo = 1
           puts foo
           instance = Object.new
           class << instance
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced in a module' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do
           foo = 1
           puts foo
           module SomeModule
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
             bar = 3
             puts bar
           end
         end
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(5)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned and unreferenced in top level' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         foo = 1
+        ^^^ Useless assignment to variable - `foo`.
         bar = 2
         puts bar
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(1)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned with operator assignment ' \
           'in top level' do
-    let(:source) do
-      'foo ||= 1'
-    end
-
     it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message).to eq(
-        'Useless assignment to variable - `foo`. Use `||` instead of `||=`.'
-      )
-      expect(cop.offenses.first.line).to eq(1)
-      expect(cop.highlights).to eq(['foo'])
+      expect_offense(<<-RUBY.strip_indent)
+        foo ||= 1
+        ^^^ Useless assignment to variable - `foo`. Use `||` instead of `||=`.
+      RUBY
     end
   end
 
   context 'when a variable is assigned multiple times ' \
           'but unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers offenses for each assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           bar = 2
           foo = 3
+          ^^^ Useless assignment to variable - `foo`.
           puts bar
         end
       RUBY
-    end
-
-    it 'registers offenses for each assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(2)
-
-      expect(cop.offenses[0].message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses[0].line).to eq(2)
-
-      expect(cop.offenses[1].message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses[1].line).to eq(4)
-
-      expect(cop.highlights).to eq(%w[foo foo])
     end
   end
 
   context 'when a referenced variable is reassigned ' \
           'but not re-referenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the non-re-referenced assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
           puts foo
           foo = 3
+          ^^^ Useless assignment to variable - `foo`.
         end
       RUBY
-    end
-
-    it 'registers an offense for the non-re-referenced assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(4)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when an unreferenced variable is reassigned ' \
           'and re-referenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the unreferenced assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           foo = 3
           puts foo
         end
       RUBY
-    end
-
-    it 'registers an offense for the unreferenced assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -345,42 +245,26 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
   end
 
   context 'when a block local variable is assigned and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers offenses for the assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         1.times do |i; foo|
           foo = 2
+          ^^^ Useless assignment to variable - `foo`.
         end
       RUBY
-    end
-
-    it 'registers offenses for the assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned in loop body and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           while true
             foo = 1
+            ^^^ Useless assignment to variable - `foo`.
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -474,97 +358,65 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a referenced variable is reassigned ' \
           'but not re-referenced in a method defined in loop' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         while true
           def some_method
             foo = 1
             puts foo
             foo = 3
+            ^^^ Useless assignment to variable - `foo`.
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(5)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable that has same name as outer scope variable ' \
           'is not referenced in a method defined in loop' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         foo = 1
 
         while foo < 100
           foo += 1
           def some_method
             foo = 1
+            ^^^ Useless assignment to variable - `foo`.
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned in single branch if ' \
           'and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           if flag
             foo = 1
+            ^^^ Useless assignment to variable - `foo`.
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a unreferenced variable is reassigned in same branch ' \
           'and referenced after the branching' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the unreferenced assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           if flag
             foo = 1
+            ^^^ Useless assignment to variable - `foo`.
             foo = 2
           end
 
           foo
         end
       RUBY
-    end
-
-    it 'registers an offense for the unreferenced assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -705,10 +557,11 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a variable is reassigned in single branch if ' \
           'and referenced in the branch' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the unreferenced assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
 
           if flag
             foo = 2
@@ -717,24 +570,16 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
         end
       RUBY
     end
-
-    it 'registers an offense for the unreferenced assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is assigned in each branch of if ' \
           'and referenced in the else branch' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the assignment in the if branch' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           if flag
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
           else
             foo = 3
             puts foo
@@ -742,104 +587,72 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
         end
       RUBY
     end
-
-    it 'registers an offense for the assignment in the if branch' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is reassigned and unreferenced in a if branch ' \
           'while the variable is referenced in the paired else branch ' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the reassignment in the if branch' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           foo = 1
 
           if flag
             puts foo
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
           else
             puts foo
           end
         end
       RUBY
     end
-
-    it 'registers an offense for the reassignment in the if branch' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context "when there's an unreferenced assignment in top level if branch " \
           'while the variable is referenced in the paired else branch' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the assignment in the if branch' do
+      expect_offense(<<-RUBY.strip_indent)
         if flag
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
         else
           puts foo
         end
       RUBY
     end
-
-    it 'registers an offense for the assignment in the if branch' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context "when there's an unreferenced reassignment in a if branch " \
           'while the variable is referenced in the paired elsif branch' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the reassignment in the if branch' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag_a, flag_b)
           foo = 1
 
           if flag_a
             puts foo
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
           elsif flag_b
             puts foo
           end
         end
       RUBY
     end
-
-    it 'registers an offense for the reassignment in the if branch' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context "when there's an unreferenced reassignment in a if branch " \
           'while the variable is referenced in a case branch ' \
           'in the paired else branch' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the reassignment in the if branch' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag_a, flag_b)
           foo = 1
 
           if flag_a
             puts foo
             foo = 2
+            ^^^ Useless assignment to variable - `foo`.
           else
             case
             when flag_b
@@ -848,15 +661,6 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
           end
         end
       RUBY
-    end
-
-    it 'registers an offense for the reassignment in the if branch' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -897,21 +701,13 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
   context 'when a variable is assigned in branch of modifier if ' \
           'that references the variable in its conditional clause' \
           'and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(flag)
           foo = 1 unless foo
+          ^^^ Useless assignment to variable - `foo`.
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -931,23 +727,15 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a unreferenced variable is reassigned ' \
           'on the left side of && and referenced after the &&' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the unreferenced assignment' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           (foo = do_something_returns_object_or_nil) && do_something
           foo
         end
       RUBY
-    end
-
-    it 'registers an offense for the unreferenced assignment' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1014,23 +802,15 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
   context 'when a variable is reassigned with binary operator ' \
            'assignment while assigning to itself in rhs ' \
            'then referenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense for the assignment in rhs' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
           foo += foo = 2
+                 ^^^ Useless assignment to variable - `foo`.
           foo
         end
       RUBY
-    end
-
-    it 'registers an offense for the assignment in rhs' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1049,69 +829,41 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a variable is assigned with ||= ' \
           'at the last expression of the scope' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = do_something_returns_object_or_nil
           foo ||= 1
+          ^^^ Useless assignment to variable - `foo`. Use `||` instead of `||=`.
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message).to eq(
-        'Useless assignment to variable - `foo`. Use `||` instead of `||=`.'
-      )
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned with ||= ' \
           'before the last expression of the scope' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = do_something_returns_object_or_nil
           foo ||= 1
+          ^^^ Useless assignment to variable - `foo`.
           some_return_value
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is assigned with multiple assignment ' \
           'and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo, bar = do_something
+               ^^^ Useless assignment to variable - `bar`. Use `_` or `_bar` as a variable name to indicate that it won't be used.
           puts foo
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message).to eq(
-        'Useless assignment to variable - `bar`. ' \
-        'Use `_` or `_bar` as a variable name ' \
-        "to indicate that it won't be used."
-      )
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['bar'])
     end
   end
 
@@ -1160,24 +912,16 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a variable is assigned ' \
           'in main body of begin with rescue but unreferenced' do
-    let(:source) do
+    it 'registers an offense' do
       <<-RUBY.strip_indent
         begin
           do_something
           foo = true
+          ^^^ Useless assignment to variable - `foo`.
         rescue
           do_anything
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1264,41 +1008,34 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a variable is reassigned multiple times in rescue ' \
           'and referenced after the begin' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         foo = false
 
         begin
           do_something
         rescue
           foo = true
+          ^^^ Useless assignment to variable - `foo`.
           foo = true
         end
 
         puts foo
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a variable is reassigned multiple times ' \
           'in rescue with ensure then referenced after the begin' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         foo = false
 
         begin
           do_something
         rescue
           foo = true
+          ^^^ Useless assignment to variable - `foo`.
           foo = true
         ensure
           do_anything
@@ -1307,41 +1044,24 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
         puts foo
       RUBY
     end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
-    end
   end
 
   context 'when a variable is reassigned multiple times ' \
           'in ensure with rescue then referenced after the begin' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         begin
           do_something
         rescue
           do_anything
         ensure
           foo = true
+          ^^^ Useless assignment to variable - `foo`.
           foo = true
         end
 
         puts foo
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(6)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1417,24 +1137,16 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a rescued error variable is wrongly tried to be referenced ' \
           'in another rescue body' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         begin
           do_something
         rescue FirstError => error
+                             ^^^^^ Useless assignment to variable - `error`.
         rescue SecondError
           p error # => nil
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `error`.')
-      expect(cop.offenses.first.line).to eq(3)
-      expect(cop.highlights).to eq(['error'])
     end
   end
 
@@ -1454,43 +1166,27 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a local variable is unreferenced ' \
           'and zero arity super is called' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(bar)
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           super
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
   context 'when a method argument is reassigned ' \
           'but not passed to super' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method(foo, bar)
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           super(bar)
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1510,21 +1206,13 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a named capture is unreferenced ' \
           'in other than top level' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-'RUBY'.strip_indent)
         def some_method
-          /(?<foo>\\w+)/ =~ 'FOO'
+          /(?<foo>\w+)/ =~ 'FOO'
+          ^^^^^^^^^^^^^ Useless assignment to variable - `foo`.
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['/(?<foo>\w+)/'])
     end
   end
 
@@ -1575,24 +1263,16 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
 
   context 'when a variable is shadowed by a block argument ' \
           'and unreferenced' do
-    let(:source) do
-      <<-RUBY.strip_indent
+    it 'registers an offense' do
+      expect_offense(<<-RUBY.strip_indent)
         def some_method
           foo = 1
+          ^^^ Useless assignment to variable - `foo`.
           1.times do |foo|
             puts foo
           end
         end
       RUBY
-    end
-
-    it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(2)
-      expect(cop.highlights).to eq(['foo'])
     end
   end
 
@@ -1711,17 +1391,11 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
   end
 
   context 'when there is only one AST node and it is unused variable' do
-    let(:source) do
-      'foo = 1'
-    end
-
     it 'registers an offense' do
-      inspect_source(source)
-      expect(cop.offenses.size).to eq(1)
-      expect(cop.offenses.first.message)
-        .to eq('Useless assignment to variable - `foo`.')
-      expect(cop.offenses.first.line).to eq(1)
-      expect(cop.highlights).to eq(['foo'])
+      expect_offense(<<-RUBY.strip_indent)
+        foo = 1
+        ^^^ Useless assignment to variable - `foo`.
+      RUBY
     end
   end
 
@@ -1740,20 +1414,12 @@ RSpec.describe RuboCop::Cop::Lint::UselessAssignment do
     end
 
     context 'and the variable is not used' do
-      let(:source) do
-        <<-RUBY.strip_indent
+      it 'registers an offense' do
+        expect_offense(<<-RUBY.strip_indent)
           some_method(foo = 1) do
+                      ^^^ Useless assignment to variable - `foo`.
           end
         RUBY
-      end
-
-      it 'registers an offense' do
-        inspect_source(source)
-        expect(cop.offenses.size).to eq(1)
-        expect(cop.offenses.first.message)
-          .to eq('Useless assignment to variable - `foo`.')
-        expect(cop.offenses.first.line).to eq(1)
-        expect(cop.highlights).to eq(['foo'])
       end
     end
   end