+++ spec/rubocop/formatter/simple_text_formatter_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Formatter::SimpleTextFormatter do
+  subject(:formatter) { described_class.new(output) }
+
   before do
     Rainbow.enabled = true
   end
-
-  subject(:formatter) { described_class.new(output) }
 
   let(:output) { StringIO.new }